[Change] separate messages controllers from routes
